Login

+++ src/main/java/com/example/Saceva2/Controller/Avvio.java
@@ -1,6 +1,7 @@
 package com.example.Saceva2.Controller;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.Saceva2.Bo.Account;
@@ -32,7 +34,40 @@ public class Avvio {// prova
 	@Autowired
 	IRepoPermesso ip;
 
-	//////////////////////////// Utente\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
+	////////////////////////////Login\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
+	@PostMapping(value = "login")
+	public String login(Account a) {
+		String esito = "";
+		System.out.println("From html username : " + a.getuName() + ", Email : "+ a.getEmail() + "pass : " + a.getPass());
+		List<Account> accounts = ia.findAll();
+		if(null == a.getEmail())
+			System.out.println();
+		if("".equals(a.getuName()) || null == a.getuName())
+			a.setuName("");
+		if("".equals(a.getEmail()) || null == a.getEmail())
+			a.setEmail("");
+		if(("".equals(a.getuName()) || null == a.getuName()) && ("".equals(a.getEmail()) || null == a.getEmail()))
+			esito = "missing UserName and Password , try again";
+		else {
+			for(int i = 0; i < accounts.size() ; i++) {
+				if((a.getuName().equals(accounts.get(i).getuName()) || a.getEmail().equals(accounts.get(i).getEmail())) 
+						&& a.getPass().equals(accounts.get(i).getPass())) {
+					System.out.println("account found on db");
+					System.out.println("From html username : " + a.getuName() + ", Email : "+ a.getEmail() + "pass : " + a.getPass());
+					System.out.println("Found on db username : " + accounts.get(i).getuName());
+					esito = "Accesso eseguito";
+					break;
+				}
+				if( i == accounts.size() - 1) {
+					System.err.println("corrispondenza non trovata");
+					esito = "Accesso negato";
+				}
+			}
+		}
+		return esito;
+	}
+
+	////////////////////////////Utente\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
 	@RequestMapping(value = "/insertUtente", method = RequestMethod.POST)
 	public List<Utente> register(Utente u, Account a, int idPermesso) throws InterruptedException {// insert and update
 		System.out.println("method of register");
@@ -61,16 +96,37 @@ public class Avvio {// prova
 
 	}
 
-	@RequestMapping(value = "/delete", method = RequestMethod.DELETE)
-	public List<Utente> delete(@RequestParam int idUtente) {// work delete account and utente
-		boolean flag = false;
+	@RequestMapping(value = "/deleteUtente", method = RequestMethod.DELETE)
+	public String delete(@RequestParam(defaultValue = "0") int idUtente, int[] listToDelete) {// work delete account and utente
 		System.out.println("method of delete");
-//		Utente u = ru.findById(idUtente);
-//		System.out.println("utente : " + u.toString());
-//		u.setAccount(a);
-//		a.setRuolo(p);
-		iu.deleteById(idUtente);
-		return iu.findAll();
+		String msg = "";
+		ArrayList<Utente> allUtenti = (ArrayList<Utente>) iu.findAll();
+		HashMap<Integer,Utente> utentiCheckToDelete = new HashMap<Integer,Utente>();
+		if(idUtente == 0 && (listToDelete.length == 0 || listToDelete == null)) {
+			msg = "sorry but idUtente or listToDelete are missing";
+		}else {
+			//check if listToDelete id exist on db
+			for(int i = 0; i < listToDelete.length; i++) {
+				Utente u = iu.findById(listToDelete[i]);
+				if(u != null) 
+					utentiCheckToDelete.put(u.getIdUtente(), u);
+			}
+			System.out.println("idUtente : " + idUtente);
+			System.out.println("List idUtenti verificati da eliminare : " + utentiCheckToDelete.toString());
+			if(utentiCheckToDelete.size() != 0 && idUtente == 0) {
+				for(int toDelete:utentiCheckToDelete.keySet()) {
+					msg = "Delte utente with id : " + toDelete;
+					iu.deleteById(toDelete);
+				}
+			}else if(idUtente != 0) {
+				if(iu.findById(idUtente) != null) {
+					msg = "Delete one Utente with id : " + idUtente;
+					iu.deleteById(idUtente);
+				}
+			}
+		}
+		
+		return msg;
 	}
 
 //	@RequestMapping(value = "tabellaUteni", method = RequestMethod.GET)
@@ -95,7 +151,7 @@ public class Avvio {// prova
 		return "tabellaUtente";
 	}
 
-	////////////////////// permesso\\\\\\\\\\\\\\\\\\\\ da provare solo la ricerca
+	//////////////////////permesso\\\\\\\\\\\\\\\\\\\\ da provare solo la ricerca
 	@PostMapping(value = "insertRuolo")
 	public List<Permesso> insertRuolo(Permesso p) {
 		System.out.println("permesso : " + p.toString());
